Removed kubevirt_vmi_migration_succeeded and kubevirt_vmi_migration_failed because there is coverage here: https://github.com/kubevirt/kubevirt/blob/main/tests/monitoring/vm_monitoring.go#L202 https://github.com/kubevirt/kubevirt/blob/main/tests/monitoring/vm_monitoring.go#L231
Moved fixtures to conftest
